requirements.txt
# Python dependencies for AI Backend Unified Infrastructure
# Install: pip install -r requirements.txt

# Configuration validation
pydantic>=2.0.0
pyyaml>=6.0

# Schema validation
jsonschema>=4.17.0

# Code quality and type checking
mypy>=1.8.0
ruff>=0.1.0

# Pre-commit hooks (optional, can also be installed globally)
pre-commit>=3.0.0
yamllint>=1.33.0
detect-secrets>=1.4.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0  # Coverage reporting
pytest-xdist>=3.3.0  # Parallel test execution
requests>=2.31.0  # HTTP requests for integration tests
pytest-playwright>=0.4.0  # Browser automation for monitoring tests
playwright>=1.40.0  # Playwright browser automation
